add new styles new add.js module
