--- index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <title>To Do List</title>
-<script defer src="main.bundle.js"></script></head>
-<body>
-  <ul class="header-container">
-    <li class="list-header">
-      <h2 class="list-title">Today's To Do</h2>
-      <div id="refresh"></div>
-    </li>
-  </ul>
-  <div class="list-container">
-  <ul class="add-new">
-    <li>
-      <form id="form" class="form">
-        <input id="new-item" placeholder="Add to your list..." type="text">
-        <button class="enter-btn" type="submit"></button>
-      </form>
-    </li>
-  </ul>
-    <ul class="list-items"></ul>
-    <div class="button-container">
-    <button type="button" class="clear-btn">Clear all completed</button>
-    </div>
-  </div>
-</body>
-</html>
