+++ src/style.css
@@ -38,6 +38,7 @@ input::placeholder {
   margin: 40px 0 0 0;
   padding: 10px;
   border: 1px solid rgb(211, 211, 211);
+  box-shadow: 0 1px 5px gray;
 }
 
 .enter-btn {
@@ -73,6 +74,7 @@ input::placeholder {
 .list-container {
   width: 50%;
   border: 1px solid rgb(211, 211, 211);
+  box-shadow: 0 5px 10px gray;
 }
 
 .add-new {
